backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-neo4j>=0.1.0
google-genai>=1.0.0
neo4j>=5.20.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
cachetools>=5.4.0
tenacity>=9.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
black>=24.0.0
isort>=5.13.0
semanticscholar>=0.8.0
sentence-transformers>=3.0.0
torch>=2.4.0
